feat: shared task-supervision utility + cache-server/cache-sdk adoption (W15)
Extract the mediator's in-tree TaskSupervisor (restart-on-failure with
capped exponential backoff + observable HealthRegistry + CancellationToken
shutdown) into a new zero-domain-dep leaf crate, affinidi-task-utils, so the
workspace shares one 'restart-and-degrade, never fail-fast' policy.

The home must be a leaf crate: a shared supervisor used by the foundational
cache-sdk cannot live in tdk-common (which depends on cache-sdk) or in the
messaging-specific mediator-common.

Consumers:
- mediator: crate::tasks::supervisor now re-exports the crate; no behaviour
  change, call sites unchanged (0.15.45 -> 0.15.46).
- cache-server: statistics task supervised (non-load-bearing) instead of a
  bare spawn+join; panic/error restarts with backoff (0.9.0 -> 0.9.1).
- cache-sdk: network task supervised, completing the restart-supervision
  deferred from W3 — panic-restart + observable network_health(); stop() now
  cancels the shutdown token (async-safe) instead of blocking_send; the
  redundant WSCommands::Exit was removed (0.8.8 -> 0.8.9). Network feature
  only; local builds unaffected.

didcomm-service's richer per-listener supervisor is intentionally left as-is
(deferred): flattening it onto the generic supervisor would regress its
connection-handle/event-broadcast features (revisit when a 2nd need appears).

Panic-injection tests per consumer; workspace build + clippy -D (default,
cache-sdk/network, mediator/fjall) + tests + fmt green.
