+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "crates/core/affinidi-crypto",
   "crates/core/affinidi-secrets-resolver",
   "crates/core/affinidi-bbs",
+  "crates/core/affinidi-task-utils",
 
   # Identity: DID resolution, authentication
   "crates/identity/affinidi-did-common",
@@ -86,6 +87,7 @@ lto = true
 # Only needed for crates that are referenced by version (not path) in other workspace members.
 [patch.crates-io]
 affinidi-secrets-resolver = { path = "crates/core/affinidi-secrets-resolver" }
+affinidi-task-utils = { path = "crates/core/affinidi-task-utils" }
 affinidi-did-resolver-cache-sdk = { path = "crates/identity/affinidi-did-resolver-cache-sdk" }
 affinidi-did-common = { path = "crates/identity/affinidi-did-common" }
 affinidi-crypto = { path = "crates/core/affinidi-crypto" }
