+++ crates/core/affinidi-task-utils/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "affinidi-task-utils"
+description = "Background-task supervision: restart-on-failure with capped backoff and an observable health registry, for long-lived tokio tasks."
+version = "0.1.0"
+edition.workspace = true
+authors.workspace = true
+readme = "README.md"
+homepage.workspace = true
+license.workspace = true
+keywords = ["ssi", "affinidi", "tokio", "supervision", "tasks"]
+repository.workspace = true
+publish.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+dashmap = "6"
+serde = { version = "1", features = ["derive"] }
+tokio = { version = "1", features = ["rt", "time", "macros"] }
+tokio-util = "0.7"
+tracing = "0.1"
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "time", "sync", "test-util"] }
