+++ crates/identity/affinidi-did-resolver-cache-server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "affinidi-did-resolver-cache-server"
-version = "0.9.0"
+version = "0.9.1"
 description = "Affinidi DID Network Cache + Resolver Service"
 edition.workspace = true
 authors.workspace = true
@@ -24,6 +24,8 @@ network = ["affinidi-did-resolver-cache-sdk/network"]
 # Affinidi Crates
 affinidi-did-resolver-cache-sdk = { version = "0.8", default-features = true, path = "../affinidi-did-resolver-cache-sdk/" }
 affinidi-did-common = "0.3"
+# Shared background-task supervision (restart-on-failure + health registry)
+affinidi-task-utils = "0.1"
 
 didwebvh-rs = "0.5"
